Sunday Update
Created most components of the experiment. Sentences are working well, and I replaced the association task with a meaning check for the novel wods. Priming.js is still a mess, trying to figure out how to remove words like "boff" and "geck" without messing up the number of testing trials there are.